move images to a folder and add header logo
